requirements.txt
alembic==1.7.6
anyio==3.5.0
apache-airflow==2.2.4
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.3
apache-airflow-providers-imap==2.2.0
apache-airflow-providers-sqlite==2.1.0
apispec==3.3.2
argcomplete==2.0.0
attrs==20.3.0
Babel==2.9.1
beautifulsoup4==4.10.0
blinker==1.4
bs4==0.0.1
cachelib==0.6.0
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
connexion==2.12.0
croniter==1.3.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
dnspython==2.2.0
docutils==0.16
email-validator==1.1.3
Flask==1.1.4
Flask-AppBuilder==3.4.4
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.19.1
gunicorn==20.1.0
h11==0.12.0
html5lib==1.1
httpcore==0.14.7
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
inflection==0.5.1
iso8601==1.0.2
itsdangerous==1.1.0
Jinja2==3.0.3
jsonschema==3.2.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
lxml==4.8.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mysql-connector==2.2.9
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pendulum==2.1.2
prison==0.2.1
psutil==5.9.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
requests==2.27.1
rfc3986==1.5.0
rich==11.2.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
unicodecsv==0.14.1
urllib3==1.26.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.13.3
WTForms==2.3.3
zipfile36==0.1.3
zipp==3.7.0
